+++ include/GafferScene/CameraTweaks.h
@@ -38,9 +38,9 @@
 #define GAFFERSCENE_CAMERATWEAKS_H
 
 #include "GafferScene/ObjectProcessor.h"
-#include "GafferScene/TweakPlug.h"
 
 #include "Gaffer/StringPlug.h"
+#include "Gaffer/TweakPlug.h"
 
 namespace GafferScene
 {
@@ -55,8 +55,8 @@ class GAFFERSCENE_API CameraTweaks : public ObjectProcessor
 
 		GAFFER_NODE_DECLARE_TYPE( GafferScene::CameraTweaks, CameraTweaksTypeId, ObjectProcessor );
 
-		GafferScene::TweaksPlug *tweaksPlug();
-		const GafferScene::TweaksPlug *tweaksPlug() const;
+		Gaffer::TweaksPlug *tweaksPlug();
+		const Gaffer::TweaksPlug *tweaksPlug() const;
 
 	protected :
