+++ include/GafferScene/EditScopeAlgo.h
@@ -38,10 +38,10 @@
 #define GAFFERSCENE_EDITSCOPEALGO_H
 
 #include "GafferScene/ScenePlug.h"
-#include "GafferScene/TweakPlug.h"
 
 #include "Gaffer/EditScope.h"
 #include "Gaffer/TransformPlug.h"
+#include "Gaffer/TweakPlug.h"
 
 #include "IECoreScene/ShaderNetwork.h"
 
@@ -109,7 +109,7 @@ GAFFERSCENE_API const Gaffer::GraphComponent *transformEditReadOnlyReason( const
 // These methods edit shader parameters for a particular location.
 
 GAFFERSCENE_API bool hasParameterEdit( const Gaffer::EditScope *scope, const ScenePlug::ScenePath &path, const std::string &attribute, const IECoreScene::ShaderNetwork::Parameter &parameter );
-GAFFERSCENE_API TweakPlug *acquireParameterEdit( Gaffer::EditScope *scope, const ScenePlug::ScenePath &path, const std::string &attribute, const IECoreScene::ShaderNetwork::Parameter &parameter, bool createIfNecessary = true );
+GAFFERSCENE_API Gaffer::TweakPlug *acquireParameterEdit( Gaffer::EditScope *scope, const ScenePlug::ScenePath &path, const std::string &attribute, const IECoreScene::ShaderNetwork::Parameter &parameter, bool createIfNecessary = true );
 GAFFERSCENE_API void removeParameterEdit( Gaffer::EditScope *scope, const ScenePlug::ScenePath &path, const std::string &attribute, const IECoreScene::ShaderNetwork::Parameter &parameter );
 GAFFERSCENE_API const Gaffer::GraphComponent *parameterEditReadOnlyReason( const Gaffer::EditScope *scope, const ScenePlug::ScenePath &path, const std::string &attribute, const IECoreScene::ShaderNetwork::Parameter &parameter );
 
@@ -119,7 +119,7 @@ GAFFERSCENE_API const Gaffer::GraphComponent *parameterEditReadOnlyReason( const
 // These methods edit attributes for a particular location.
 
 GAFFERSCENE_API bool hasAttributeEdit( const Gaffer::EditScope *scope, const ScenePlug::ScenePath &path, const std::string &attribute );
-GAFFERSCENE_API TweakPlug *acquireAttributeEdit( Gaffer::EditScope *scope, const ScenePlug::ScenePath &path, const std::string &attribute, bool createIfNecessary = true );
+GAFFERSCENE_API Gaffer::TweakPlug *acquireAttributeEdit( Gaffer::EditScope *scope, const ScenePlug::ScenePath &path, const std::string &attribute, bool createIfNecessary = true );
 GAFFERSCENE_API void removeAttributeEdit( Gaffer::EditScope *scope, const ScenePlug::ScenePath &path, const std::string &attribute );
 GAFFERSCENE_API const Gaffer::GraphComponent *attributeEditReadOnlyReason( const Gaffer::EditScope *scope, const ScenePlug::ScenePath &path, const std::string &attribute );
