+++ python/GafferArnold/ArnoldTextureBake.py
@@ -204,14 +204,14 @@ class ArnoldTextureBake( GafferDispatch.TaskNode ) :
 
 			self["__cameraTweaks"] = GafferScene.CameraTweaks()
 			self["__cameraTweaks"]["in"].setInput( self["__camera"]["out"] )
-			self["__cameraTweaks"]["tweaks"]["projection"] = GafferScene.TweakPlug( "projection", "uv_camera" )
-			self["__cameraTweaks"]["tweaks"]["resolution"] = GafferScene.TweakPlug( "resolution", imath.V2i( 0 ) )
-			self["__cameraTweaks"]["tweaks"]["u_offset"] = GafferScene.TweakPlug( "u_offset", 0.0 )
-			self["__cameraTweaks"]["tweaks"]["v_offset"] = GafferScene.TweakPlug( "v_offset", 0.0 )
-			self["__cameraTweaks"]["tweaks"]["mesh"] = GafferScene.TweakPlug( "mesh", "" )
-			self["__cameraTweaks"]["tweaks"]["uv_set"] = GafferScene.TweakPlug( "uv_set", "" )
-			self["__cameraTweaks"]["tweaks"]["extend_edges"] = GafferScene.TweakPlug( "extend_edges", False )
-			self["__cameraTweaks"]["tweaks"]["offset"] = GafferScene.TweakPlug( "offset", 0.1 )
+			self["__cameraTweaks"]["tweaks"]["projection"] = Gaffer.TweakPlug( "projection", "uv_camera" )
+			self["__cameraTweaks"]["tweaks"]["resolution"] = Gaffer.TweakPlug( "resolution", imath.V2i( 0 ) )
+			self["__cameraTweaks"]["tweaks"]["u_offset"] = Gaffer.TweakPlug( "u_offset", 0.0 )
+			self["__cameraTweaks"]["tweaks"]["v_offset"] = Gaffer.TweakPlug( "v_offset", 0.0 )
+			self["__cameraTweaks"]["tweaks"]["mesh"] = Gaffer.TweakPlug( "mesh", "" )
+			self["__cameraTweaks"]["tweaks"]["uv_set"] = Gaffer.TweakPlug( "uv_set", "" )
+			self["__cameraTweaks"]["tweaks"]["extend_edges"] = Gaffer.TweakPlug( "extend_edges", False )
+			self["__cameraTweaks"]["tweaks"]["offset"] = Gaffer.TweakPlug( "offset", 0.1 )
 
 			self["__cameraTweaks"]["tweaks"]["offset"]["value"].setInput( self["normalOffset"] )
