+++ python/GafferSceneUI/__init__.py
@@ -47,7 +47,6 @@ from .PrimitiveInspector import PrimitiveInspector
 from .UVInspector import UVInspector
 from .FilterPlugValueWidget import FilterPlugValueWidget
 from .ScenePathPlugValueWidget import ScenePathPlugValueWidget
-from .TweakPlugValueWidget import TweakPlugValueWidget
 from .LightEditor import LightEditor
 from . import SceneHistoryUI
 from . import EditScopeUI
