+++ src/GafferSceneModule/TweaksBinding.cpp
@@ -42,7 +42,6 @@
 #include "GafferScene/AttributeTweaks.h"
 #include "GafferScene/ShaderTweaks.h"
 #include "GafferScene/CameraTweaks.h"
-#include "GafferScene/TweakPlug.h"
 
 #include "GafferBindings/DependencyNodeBinding.h"
 #include "GafferBindings/PlugBinding.h"
@@ -54,140 +53,9 @@ using namespace Gaffer;
 using namespace GafferBindings;
 using namespace GafferScene;
 
-namespace
-{
-
-TweakPlugPtr constructUsingData( const std::string &tweakName, IECore::ConstDataPtr tweakValue, TweakPlug::Mode mode, bool enabled )
-{
-	return new TweakPlug( tweakName, tweakValue.get(), mode, enabled );
-}
-
-bool applyTweak( const TweakPlug &plug, IECore::CompoundData &parameters, TweakPlug::MissingMode missingMode )
-{
-	IECorePython::ScopedGILRelease gilRelease;
-	return plug.applyTweak( &parameters, missingMode );
-}
-
-bool applyTweaks( const Plug &tweaksPlug, IECoreScene::ShaderNetwork &shaderNetwork, TweakPlug::MissingMode missingMode )
-{
-	IECorePython::ScopedGILRelease gilRelease;
-	return TweakPlug::applyTweaks( &tweaksPlug, &shaderNetwork, missingMode );
-}
-
-bool applyTweaksToParameters( const TweaksPlug &tweaksPlug, IECore::CompoundData &parameters, TweakPlug::MissingMode missingMode )
-{
-	IECorePython::ScopedGILRelease gilRelease;
-	return tweaksPlug.applyTweaks( &parameters, missingMode );
-}
-
-class TweakPlugSerialiser : public ValuePlugSerialiser
-{
-	bool childNeedsConstruction( const Gaffer::GraphComponent *child, const Serialisation &serialisation ) const override
-	{
-		return false;
-	}
-
-	std::string constructor( const Gaffer::GraphComponent *graphComponent, Serialisation &serialisation ) const override
-	{
-		auto tweaksPlug = static_cast<const TweakPlug *>( graphComponent );
-
-		const Serialiser *valuePlugSerialiser = Serialisation::acquireSerialiser( tweaksPlug->valuePlug() );
-		std::string result = ValuePlugSerialiser::constructor( graphComponent, serialisation );
-
-		// Pass the value plug into the constructor directly so that there's
-		// never a moment in which the TweakPlug is in an invalid state.
-		result = boost::algorithm::replace_first_copy(
-			result,
-			"TweakPlug(",
-			"TweakPlug( " + valuePlugSerialiser->constructor( tweaksPlug->valuePlug(), serialisation ) + ","
-		);
-
-		return result;
-	}
-};
-
-} // namespace
-
 void GafferSceneModule::bindTweaks()
 {
 	DependencyNodeClass<ShaderTweaks>();
 	DependencyNodeClass<CameraTweaks>();
 	DependencyNodeClass<AttributeTweaks>();
-
-	PlugClass<TweakPlug> tweakPlugClass;
-
-	{
-		scope tweakPlugScope = tweakPlugClass;
-
-		enum_<TweakPlug::Mode>( "Mode" )
-			.value( "Replace", TweakPlug::Replace )
-			.value( "Add", TweakPlug::Add )
-			.value( "Subtract", TweakPlug::Subtract )
-			.value( "Multiply", TweakPlug::Multiply )
-			.value( "Remove", TweakPlug::Remove )
-			.value( "Create", TweakPlug::Create )
-		;
-
-		enum_<TweakPlug::MissingMode>( "MissingMode" )
-			.value( "Ignore", TweakPlug::MissingMode::Ignore )
-			.value( "Error", TweakPlug::MissingMode::Error )
-			.value( "IgnoreOrReplace", TweakPlug::MissingMode::IgnoreOrReplace )
-		;
-	}
-
-	tweakPlugClass
-		.def(
-			init<ValuePlug *, const char *, Plug::Direction, unsigned>(
-				(
-					boost::python::arg_( "valuePlug" ),
-					boost::python::arg_( "name" )=GraphComponent::defaultName<TweakPlug>(),
-					boost::python::arg_( "direction" )=Plug::In,
-					boost::python::arg_( "flags" )=Plug::Default
-				)
-			)
-		)
-		.def(
-			"__init__",
-			make_constructor(
-				constructUsingData,
-				default_call_policies(),
-				(
-					boost::python::arg_( "tweakName" ),
-					boost::python::arg_( "valuePlug" ),
-					arg( "mode" ) = TweakPlug::Replace,
-					boost::python::arg_( "enabled" )=true
-				)
-			)
-		)
-		.def(
-			init<const std::string &, const ValuePlugPtr, TweakPlug::Mode, bool>(
-				(
-					boost::python::arg_( "tweakName" ),
-					boost::python::arg_( "value" ),
-					arg( "mode" ) = TweakPlug::Replace,
-					boost::python::arg_( "enabled" )=true
-				)
-			)
-		)
-		.def( "applyTweak", &applyTweak, ( arg( "parameters" ), arg( "missingMode" ) = TweakPlug::MissingMode::Error ) )
-		.def( "applyTweaks", &applyTweaks, ( arg( "shaderNetwork" ), arg( "missingMode" ) = TweakPlug::MissingMode::Error ) )
-		.staticmethod( "applyTweaks" )
-	;
-
-	Serialisation::registerSerialiser( TweakPlug::staticTypeId(), new TweakPlugSerialiser );
-
-	PlugClass<TweaksPlug>()
-		.def(
-			init<const std::string &, Plug::Direction, unsigned>(
-				(
-					boost::python::arg_( "name" )=GraphComponent::defaultName<TweaksPlug>(),
-					boost::python::arg_( "direction" )=Plug::In,
-					boost::python::arg_( "flags" )=Plug::Default
-				)
-			)
-		)
-		.def( "applyTweaks", &applyTweaks, ( arg( "shaderNetwork" ), arg( "missingMode" ) = TweakPlug::MissingMode::Error ) )
-		.def( "applyTweaks", &applyTweaksToParameters, ( arg( "parameters" ), arg( "missingMode" ) = TweakPlug::MissingMode::Error ) )
-	;
-
 }
